댓글, 해시태그 추가 및 작성, 수정, 삭제 권한 설정

+++ blog/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render, redirect
 from django.views import View
-from .models import Post, Category
-from .forms import PostForm
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth import get_user_model
+from django.http import HttpResponseForbidden
+from .models import Post, Category, Comment, HashTag
+from .forms import PostForm, CommentForm, HashTagForm
 
 # Create your views here.
 
@@ -21,14 +24,23 @@ class Index(View):
 class DetailView(View):
     def get(self, request, pk):
         post = Post.objects.get(pk=pk)
+        comments = Comment.objects.select_related('post').filter(post=post)
+        hashtags = HashTag.objects.select_related('post').filter(post=post)
+        comment_form = CommentForm()
+        hashtag_form = HashTagForm()
+
         context = {
             'title': 'Blog',
             'post': post,
+            'comments': comments,
+            'hashtags': hashtags,
+            'comment_form': comment_form,
+            'hashtag_form': hashtag_form,
         }
         return render(request, 'blog/post_detail.html', context)
 
 
-class Write(View):
+class Write(LoginRequiredMixin, View):
     def get(self, request):
         form = PostForm()
         context = {
@@ -41,6 +53,7 @@ class Write(View):
         form = PostForm(request.POST)
         if form.is_valid():
             post = form.save(commit=False)
+            post.writer = request.user
             post.save()
             return redirect('blog:list')
         form.add_error(None, '폼이 유효하지 않습니다.')
@@ -54,35 +67,42 @@ class Write(View):
 class Update(View):
     def get(self, request, pk):
         post = Post.objects.get(pk=pk)
-        form = PostForm(initial={'title': post.title, 'content': post.content})
-        context = {
-            'title': 'Blog',
-            'post': post,
-            'form': form,
-        }
-        return render(request, 'blog/post_edit.html', context)
+        if request.user == post.writer:
+            form = PostForm(
+                initial={'title': post.title, 'content': post.content})
+            context = {
+                'title': 'Blog',
+                'post': post,
+                'form': form,
+            }
+            return render(request, 'blog/post_edit.html', context)
+        return HttpResponseForbidden("수정 권한이 없습니다.")
 
     def post(self, request, pk):
         post = Post.objects.get(pk=pk)
         form = PostForm(request.POST)
-        if form.is_valid():
-            post.title = form.cleaned_data['title']
-            post.content = form.cleaned_data['content']
-            post.save()
-            return redirect('blog:detail', pk=pk)
-        form.add_error(None, '폼이 유효하지 않습니다.')
-        context = {
-            'title': 'Blog',
-            'form': form,
-        }
-        return render(request, 'blog/post_edit.html', context)
+        if request.user == post.writer:
+            if form.is_valid():
+                post.title = form.cleaned_data['title']
+                post.content = form.cleaned_data['content']
+                post.save()
+                return redirect('blog:detail', pk=pk)
+            form.add_error(None, '폼이 유효하지 않습니다.')
+            context = {
+                'title': 'Blog',
+                'form': form,
+            }
+            return render(request, 'blog/post_edit.html', context)
+        return HttpResponseForbidden("수정 권한이 없습니다.")
 
 
 class Delete(View):
     def post(self, request, pk):
         post = Post.objects.get(pk=pk)
-        post.delete()
-        return redirect('blog:list')
+        if request.user == post.writer:
+            post.delete()
+            return redirect('blog:list')
+        return HttpResponseForbidden("삭제 권한이 없습니다.")
 
 
 class CategoryView(View):
@@ -95,3 +115,75 @@ class CategoryView(View):
             'categories': Category.objects.all(),
         }
         return render(request, 'blog/post_list.html', context)
+
+
+# ## 댓글
+class CommentWrite(LoginRequiredMixin, View):
+    def post(self, request, pk):
+        form = CommentForm(request.POST)
+        post = Post.objects.get(pk=pk)
+
+        if form.is_valid():
+            content = form.cleaned_data['content']
+            writer = request.user
+            comment = Comment.objects.create(
+                post=post, content=content, writer=writer)
+            return redirect('blog:detail', pk=pk)
+
+        hashtag_form = HashTagForm()
+        context = {
+            'title': 'Blog',
+            'post': post,
+            'comments': post.comment_set.all(),
+            'hashtags': post.hashtag_set.all(),
+            'comment_form': form,
+            'hashtag_form': hashtag_form
+        }
+        return render(request, 'blog/post_detail.html', context)
+
+
+class CommentDelete(View):
+    def post(self, request, pk):
+        comment = Comment.objects.get(pk=pk)
+        post_id = comment.post.id
+        if request.user == comment.writer or request.user == comment.post.writer:
+            comment.delete()
+            return redirect('blog:detail', pk=post_id)
+        return HttpResponseForbidden("삭제 권한이 없습니다.")
+
+
+# ## 해시태그
+class HashTagWrite(View):
+    def post(self, request, pk):
+        form = HashTagForm(request.POST)
+        post = Post.objects.get(pk=pk)
+
+        if request.user == post.writer:
+            if form.is_valid():
+                name = form.cleaned_data['name']
+                writer = request.user
+                hashtag = HashTag.objects.create(
+                    post=post, name=name, writer=writer)
+                return redirect('blog:detail', pk=pk)
+
+            comment_form = CommentForm()
+            context = {
+                'title': 'Blog',
+                'post': post,
+                'comments': post.comment_set.all(),
+                'hashtags': post.hashtag_set.all(),
+                'comment_form': comment_form,
+                'hashtag_form': form
+            }
+            return render(request, 'blog/post_detail.html', context)
+        return HttpResponseForbidden("태그 작성은 글 작성자만 가능합니다.")
+
+
+class HashTagDelete(View):
+    def post(self, request, pk):
+        hashtag = HashTag.objects.get(pk=pk)
+        post_id = hashtag.post.id
+        if request.user == hashtag.writer:
+            hashtag.delete()
+            return redirect('blog:detail', pk=post_id)
+        return HttpResponseForbidden("삭제 권한이 없습니다.")
